% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_classifier.R
\name{evaluate_classifier}
\alias{evaluate_classifier}
\title{Evaluates a trained classification model using accuracy and confusion matrix.}
\usage{
evaluate_classifier(model, testdata)
}
\arguments{
\item{model}{Trained model object (from train_classifier).}

\item{testdata}{Test data for evaluation.}
}
\value{
List with evaluation metrics: accuracy and confusion matrix.
}
\description{
Evaluates a trained classification model using accuracy and confusion matrix.
}
\examples{
model <- train_classifier(iris, Species ~ ., method = "rf")
evaluate_classifier(model, testdata = iris[-1, ])
}
