% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_classifier.R
\name{predict_classifier}
\alias{predict_classifier}
\title{Predicts using a trained Random Forest classification model.}
\usage{
predict_classifier(model, new_data)
}
\arguments{
\item{model}{A trained random forest model object (from train_classifier).}

\item{new_data}{New data to predict on.}
}
\value{
Predicted class labels.
}
\description{
Predicts using a trained Random Forest classification model.
}
\examples{
# Train a model using iris dataset
model <- train_classifier(iris, Species ~ ., method = "rf")

# Predict using new data (first 5 rows of iris dataset)
new_data <- iris[1:5, ]
predicted <- predict_classifier(model, new_data)

}
